app-backend/requirements.txt
# Backend Dependencies
Django==5.0.3
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-csp==3.7
Pillow
python-dotenv==1.0.1
psycopg2-binary
django-filter==24.1
django-storages==1.14.2
boto3==1.34.34
reportlab==4.1.0
django-allauth==0.60.1
django-rest-auth==0.9.5
djangorestframework-simplejwt==5.3.1
django-cleanup==8.0.0
django-environ==0.11.2
gunicorn==21.2.0
whitenoise==6.6.0
Faker==19.13.0
bcrypt==4.1.2
cryptography==42.0.5

# Development Dependencies
pytest==8.0.2
pytest-django==4.7.0
coverage==7.4.1
black==24.2.0
flake8==7.0.0
isort==5.13.2
mypy==1.8.0
pylint-django==2.5.5
pylint==3.0.3

# Frontend Dependencies (for reference)
# Note: These are managed by npm/yarn, but listed here for documentation
# next==15.2.1
# react==19.0.0
# react-dom==19.0.0
# lucide-react==0.477.0
# tailwindcss==4.0.0
# eslint==9.0.0 